requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
beautifulsoup4==4.9.1
boto3==1.14.23
botocore==1.17.23
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
Django==3.0.8
django-cors-headers==3.4.0
django-extensions==3.0.9
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-extra-fields==3.0.3
drf-writable-nested==0.6.0
drf-yasg==1.17.1
flake8==3.8.3
google==3.0.0
google-api-core==1.22.2
google-api-python-client==1.12.2
google-auth==1.21.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
protobuf==3.13.0
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydot==1.4.1
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.1
pytest-django==3.9.0
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sendgrid==6.4.3
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
